Support table prefixes for 1.3.8 upgrade in zc_installer
Prior to ZC 1.5.5, the zc_installer included code to process the upgrade
script from ZC 1.3.8 to 1.3.9; however, in the redevelopment for ZC 1.5.5,
operations did not include processing SQL commands such as only DROP TABLE
though DROP TABLE IF EXISTS was supported.  This code was replaced and
expanded to support DROP TABLE and DROP TABLE IF EXISTS.

Also the code did not support the SQL command of RENAME TABLE.

To support the revised code the ZC 1.3.9 sql (upgrades 1.3.8 to 1.3.9)
required minor modification as well.

No other related/earlier SQL was identified as needing this change.

+++ zc_install/includes/classes/class.zcDatabaseInstaller.php
@@ -37,7 +37,7 @@ class zcDatabaseInstaller
     $this->dieOnErrors = isset($options['dieOnErrors']) ? (bool)$options['dieOnErrors'] : FALSE;
     $this->errors = array();
     $this->basicParseStrings = array(
-    'DROP TABLE IF EXISTS ',
+    'DROP TABLE ',
     'CREATE TABLE ',
     'REPLACE INTO ',
     'INSERT INTO ',
@@ -163,9 +163,23 @@ class zcDatabaseInstaller
       error_log("MySQL error " . $this->db->error_number . " encountered during zc_install:\n" . $this->db->error_text . "\n" . $this->line . "\n---------------\n\n");
     }
   }
-  public function parserDropTableIfExists ()
+  public function parserDropTable()
   {
-    $this->line = 'DROP TABLE IF EXISTS ' . $this->dbPrefix . substr($this->line, 21);
+    $table = (strtoupper($this->lineSplit[2].' '.$this->lineSplit[3]) == 'IF EXISTS') ? $this->lineSplit[4] : $this->lineSplit[2];
+
+    if (!$this->tableExists($table) && (strtoupper($this->lineSplit[2].' '.$this->lineSplit[3]) != 'IF EXISTS'))
+    {
+      $result = sprintf(REASON_TABLE_NOT_FOUND, $table).' CHECK PREFIXES!';
+      $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+      $this->ignoreLine = true;
+    } else {
+      if (strtoupper($this->lineSplit[2].' '.$this->lineSplit[3]) != 'IF EXISTS')
+      {
+        $this->line = 'DROP TABLE ' . $this->dbPrefix . substr($this->line, 11);
+      } else {
+        $this->line = 'DROP TABLE IF EXISTS ' . $this->dbPrefix . substr($this->line, 21);
+      }
+    }
   }
   public function parserCreateTable()
   {
@@ -325,6 +339,24 @@ class zcDatabaseInstaller
             break;
           default:
             // Do nothing
+      }
+    }
+  }
+  public function parserRenameTable()
+  {
+    if (!$this->tableExists($this->lineSplit[2])) 
+    {
+      if (!isset($result)) $result = sprintf(REASON_TABLE_NOT_FOUND, $table).' CHECK PREFIXES!';
+      $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+      $this->ignoreLine = true;
+    } else {
+      if ($this->tableExists($this->lineSplit[4]))
+      {
+        if (!isset($result)) $result = sprintf(REASON_TABLE_ALREADY_EXISTS, $table);
+        $this->writeUpgradeExceptions($this->line, $result, $this->fileName);
+        $this->ignoreLine = true;
+      } else {
+        $this->line = 'RENAME TABLE ' . $this->dbPrefix . $this->lineSplit[2] . ' TO ' . $this->dbPrefix . substr($this->line, (13 + strlen($this->lineSplit[2]) + 4));
       }
     }
   }
